+++ internal/config/dot_env.go
@@ -8,16 +8,26 @@ import (
 	"github.com/subosito/gotenv"
 )
 
-// overrideEnv forcefully overrides ENV variables through the supplied .env file.
+// os.SetEnv func signature
+type envSetter = func(key string, value string) error
+
+// DotEnvTryLoad forcefully overrides ENV variables through **a maybe available** .env file.
 //
-// This mechanism should only be used **locally** to easily inject (gitignored)
-// secrets into your ENV. Non-existing .env files are actually the best case.
-// Thus, this function will always remain silent if a .env file does not exist!
-//
+// This function will always remain silent if a .env file does not exist!
 // If we successfully apply an ENV file, we will log a warning.
 // If there are any other errors, we will panic!
-func overrideEnv(absolutePathToEnvFile string) {
-	err := gotenv.OverLoad(absolutePathToEnvFile)
+//
+// This mechanism should only be used **locally** to easily inject (gitignored)
+// secrets into your ENV. Non-existing .env files are actually the **best case**.
+//
+// When running normally (not within tests):
+// DotEnvTryLoad("/path/tp/my.env.local", os.SetEnv)
+//
+// For tests (and autoreset) use t.Setenv:
+// DotEnvTryLoad("/path/to/my.env.test.local", func(k string, v string) error { t.Setenv(k, v); return nil })
+func DotEnvTryLoad(absolutePathToEnvFile string, setEnvFn envSetter) {
+	err := DotEnvLoad(absolutePathToEnvFile, setEnvFn)
+
 	if err != nil {
 		if !errors.Is(err, os.ErrNotExist) {
 			log.Panic().Err(err).Str("envFile", absolutePathToEnvFile).Msg(".env parse error!")
@@ -27,13 +37,17 @@ func overrideEnv(absolutePathToEnvFile string) {
 	}
 }
 
-// SetEnvFromFile loads a dotenv file and executes the passed setEnvFn per ENV key/value.
-// When running in test and you simply want to override the current ENV, simply pass:
+// DotEnvLoad forcefully overrides ENV variables through the supplied .env file.
 //
-// SetEnvFromFile("/path/tp/my.test.env.local", t.SetEnv)
+// This mechanism should only be used **locally** to easily inject (gitignored)
+// secrets into your ENV.
 //
-// This ensures the ENV vars will reset to their original state after your test is finished.
-func SetEnvFromFile(absolutePathToEnvFile string, setEnvFn func(key string, val string)) error {
+// When running normally (not within tests):
+// DotEnvLoad("/path/to/my.env.local", os.SetEnv)
+//
+// For tests (and ENV var autoreset) use t.Setenv:
+// DotEnvLoad("/path/to/my.env.test.local", func(k string, v string) error { t.Setenv(k, v); return nil })
+func DotEnvLoad(absolutePathToEnvFile string, setEnvFn envSetter) error {
 
 	file, err := os.Open(absolutePathToEnvFile)
 
@@ -49,11 +63,10 @@ func SetEnvFromFile(absolutePathToEnvFile string, setEnvFn func(key string, val 
 		return err
 	}
 
-	for key, val := range envs {
-		// if err := setEnvFn(key, val); err != nil {
-		// 	return err
-		// }
-		setEnvFn(key, val)
+	for key, value := range envs {
+		if err := setEnvFn(key, value); err != nil {
+			return err
+		}
 	}
 
 	return nil