+++ internal/config/dot_env_test.go
@@ -1,10 +1,11 @@
-package config
+package config_test
 
 import (
 	"os"
 	"path/filepath"
 	"testing"
 
+	"allaboutapps.dev/aw/go-starter/internal/config"
 	"allaboutapps.dev/aw/go-starter/internal/util"
 	"github.com/stretchr/testify/assert"
 )
@@ -14,13 +15,19 @@ func TestDotEnvOverride(t *testing.T) {
 
 	orgPsqlUser := os.Getenv("PSQL_USER")
 
-	overrideEnv(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env1.local"))
+	config.DotEnvTryLoad(
+		filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env1.local"),
+		func(k string, v string) error { t.Setenv(k, v); return nil })
+
 	assert.Equal(t, "yes", os.Getenv("IS_THIS_A_TEST_ENV"))
 	assert.Equal(t, "dotenv_override_psql_user", os.Getenv("PSQL_USER"))
 	assert.Equal(t, orgPsqlUser, os.Getenv("ORIGINAL_PSQL_USER"))
 
 	// override works as expected?
-	overrideEnv(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env2.local"))
+	config.DotEnvTryLoad(
+		filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env2.local"),
+		func(k string, v string) error { t.Setenv(k, v); return nil })
+
 	assert.Equal(t, "yes still", os.Getenv("IS_THIS_A_TEST_ENV"))
 	assert.NotEqual(t, "dotenv_override_psql_user", os.Getenv("PSQL_USER"))
 	assert.Equal(t, orgPsqlUser, os.Getenv("PSQL_USER"), "Reset to original does not work!")
@@ -28,13 +35,21 @@ func TestDotEnvOverride(t *testing.T) {
 
 func TestNoopEnvNotFound(t *testing.T) {
 	assert.NotPanics(t, assert.PanicTestFunc(func() {
-		overrideEnv(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.does.not.exist"))
+
+		config.DotEnvTryLoad(
+			filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.does.not.exist"),
+			func(k string, v string) error { t.Setenv(k, v); return nil })
+
 	}), "does not panic on file inexistance")
 }
 
 func TestEmptyEnv(t *testing.T) {
 	assert.NotPanics(t, assert.PanicTestFunc(func() {
-		overrideEnv(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.local.sample"))
+
+		config.DotEnvTryLoad(
+			filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.local.sample"),
+			func(k string, v string) error { t.Setenv(k, v); return nil })
+
 	}), "does not panic on file inexistance")
 
 	assert.Empty(t, os.Getenv("EMPTY_VARIABLE_INIT"), "should be empty")
@@ -42,8 +57,10 @@ func TestEmptyEnv(t *testing.T) {
 
 func TestPanicsOnEnvMalform(t *testing.T) {
 	assert.Panics(t, assert.PanicTestFunc(func() {
-		overrideEnv(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.local.malformed"))
-	}), "does panic on file malform")
 
-	SetEnvFromFile(filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.does.not.exist"), t.Setenv)
+		config.DotEnvTryLoad(
+			filepath.Join(util.GetProjectRootDir(), "/internal/config/testdata/.env.local.malformed"),
+			func(k string, v string) error { t.Setenv(k, v); return nil })
+
+	}), "does panic on file malform")
 }